Changed recipe, search,
I changed screens and navigation for recipe and ingredient search screen
