+++ src/screens/IngredientSearch/styles.js
@@ -1,0 +1,82 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "flex-start",
+    backgroundColor: "#F0EFEF",
+  },
+  searchContainer: {
+    flexDirection: "row",
+    alignItems: "center",
+    marginBottom: 20,
+    marginTop: 20,
+    width: "100%",
+  },
+  searchInputContainer: {
+    flex: 1,
+    height: 40,
+    backgroundColor: "#fff",
+    borderTopWidth: 0,
+    borderBottomWidth: 0,
+  },
+  searchInput: {
+    fontSize: 16,
+    color: "#000",
+  },
+  searchButton: {
+    backgroundColor: "#0099ff",
+    borderRadius: 5,
+    paddingVertical: 10,
+    paddingHorizontal: 20,
+  },
+  searchButtonText: {
+    color: "#fff",
+    fontSize: 16,
+    fontWeight: "bold",
+  },
+  carouselContainer: {
+    flexDirection: "row",
+    flexWrap: "wrap",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  slide: {
+    width: "48%",
+    marginVertical: 10,
+    backgroundColor: "#fff",
+    borderRadius: 5,
+    padding: 20,
+    shadowColor: "#000",
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+    elevation: 5,
+  },
+  title: {
+    fontSize: 20,
+    fontWeight: "bold",
+    marginBottom: 10,
+  },
+  photo: {
+    width: 200,
+    height: 200,
+    marginRight: 10,
+  },
+  rowContainer: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    width: "100%",
+    paddingHorizontal: 10,
+    marginBottom: 10,
+  },
+  rowItem: {
+    width: "48%",
+  },
+});
+
+export default styles;
