+++ src/screens/Search/styles.js
@@ -3,79 +3,114 @@ import { StyleSheet } from "react-native";
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    alignItems: "center",
-    justifyContent: "flex-start",
     backgroundColor: "#F0EFEF",
   },
-  searchContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    marginBottom: 20,
-    marginTop: 20,
-    width: "100%",
-  },
-  searchInputContainer: {
+  recipe: {
     flex: 1,
-    height: 40,
-    backgroundColor: "#fff",
-    borderTopWidth: 0,
-    borderBottomWidth: 0,
-  },
-  searchInput: {
-    fontSize: 16,
-    color: "#000",
-  },
-  searchButton: {
-    backgroundColor: "#0099ff",
-    borderRadius: 5,
-    paddingVertical: 10,
-    paddingHorizontal: 20,
-  },
-  searchButtonText: {
-    color: "#fff",
-    fontSize: 16,
-    fontWeight: "bold",
-  },
-  carouselContainer: {
-    flexDirection: "row",
-    flexWrap: "wrap",
-    justifyContent: "space-between",
-    alignItems: "center",
-  },
-  slide: {
-    width: "48%",
+    flexDirection: "column",
+    marginHorizontal: 20,
     marginVertical: 10,
-    backgroundColor: "#fff",
-    borderRadius: 5,
+    backgroundColor: "white",
+    borderRadius: 10,
     padding: 20,
     shadowColor: "#000",
     shadowOffset: {
       width: 0,
-      height: 2,
+      height: 1,
     },
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-    elevation: 5,
+    shadowOpacity: 0.22,
+    shadowRadius: 2.22,
+
+    elevation: 3,
   },
+  imageContainer: {
+    alignItems: "center",
+    marginBottom: 10,
+  },
+  image: {
+    width: "100%",
+    height: 200,
+    borderRadius: 10,
+  },
+  textContainer: {},
   title: {
     fontSize: 20,
     fontWeight: "bold",
+    marginBottom: 5,
+  },
+  info: {
+    fontSize: 16,
+    marginBottom: 5,
+  },
+  summary: {
+    fontSize: 16,
+    marginTop: 10,
+    marginBottom: 5,
+  },
+  ingredientsTitle: {
+    fontSize: 18,
+    fontWeight: "bold",
+    marginTop: 10,
+    marginBottom: 5,
+  },
+  ingredients: {
+    fontSize: 16,
     marginBottom: 10,
   },
-  photo: {
-    width: 200,
-    height: 200,
-    marginRight: 10,
+  instructionsTitle: {
+    fontSize: 18,
+    fontWeight: "bold",
+    marginBottom: 5,
   },
-  rowContainer: {
+  instructions: {
+    fontSize: 16,
+    marginBottom: 10,
+  },
+  addButton: {
+    backgroundColor: "#22aa88",
+    borderRadius: 5,
+    padding: 10,
+    marginTop: 10,
+    alignSelf: "flex-start",
+  },
+  buttonContainer: {
+    alignSelf: "flex-end",
+    marginTop: 10,
+    backgroundColor: "#B0B4B33A",
+    borderRadius: 10,
     flexDirection: "row",
     justifyContent: "space-between",
-    width: "100%",
-    paddingHorizontal: 10,
-    marginBottom: 10,
+    padding: 10,
+  },  
+  addButtonText: {
+    color: "#fff",
+    fontSize: 16,
+    fontWeight: "bold",
   },
-  rowItem: {
-    width: "48%",
+  filterButtonText: {
+    color: "white",
+    fontSize: 16,
+    fontWeight: "bold",
+    padding: 10,
+    borderRadius: 20,
+    marginRight: 10,
+    marginLeft: 10,
+  },
+  filterButton:{
+    backgroundColor: "#A8DDA8",
+    borderRadius: 20,
+    marginRight: 10,
+    marginLeft: 10,
+  },
+  searchBarContainer: {
+    backgroundColor: "transparent",
+    borderTopWidth: 0,
+    borderBottomWidth: 0,
+    paddingHorizontal: 8,
+  },
+
+  searchBarInputContainer: {
+    backgroundColor: "#f0f0f0",
   },
 });
